merge 1.2.0 branch

+++ src/main/java/com/zcunsoft/clklog/api/entity/clickhouse/VisitorSummarybydate.java
@@ -7,45 +7,80 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import java.sql.Timestamp;
 
+/**
+ * 访客概要情况.
+ */
 @Entity(name = "visitor_summary_bydate")
 @Data
 public class VisitorSummarybydate {
 
+    /**
+     * 统计日期.
+     */
     @Id
     @Column
     Timestamp statDate;
 
+    /**
+     * 渠道.
+     */
     @Id
     @Column
     String lib;
 
-
+    /**
+     * 是否首日访问.
+     */
     @Id
     @Column
     String isFirstDay;
 
+    /**
+     * 项目编码.
+     */
     @Id
     @Column
     String projectName;
+
+    /**
+     * 浏览量.
+     */
     @Column
-    Integer pv;
+    Long pv;
 
+    /**
+     * 访问次数.
+     */
     @Column
-    Integer visitCount;
+    Long visitCount;
 
+    /**
+     * 访客数.
+     */
     @Column
-    Integer uv;
+    Long uv;
 
+    /**
+     * Ip数.
+     */
     @Column
-    Integer ipCount;
+    Long ipCount;
 
+    /**
+     * 访问时长
+     */
     @Column
-    Integer visitTime;
+    Long visitTime;
 
+    /**
+     * 跳出次数.
+     */
     @Column
-    Integer bounceCount;
+    Long bounceCount;
 
-
+    /**
+     * 更新时间.
+     */
     @Column
     Timestamp updateTime;
 }
